refactor: rename some model files and remove unnecessary commas
