+++ domain/src/main/java/com/baghdad/domain/model/savedList/SavedListDetails.kt
@@ -1,9 +1,9 @@
 package com.baghdad.domain.model.savedList
 
-import com.baghdad.domain.model.PagedResult
+import com.baghdad.domain.model.pagination.PagedResult
 import com.baghdad.entity.savedList.SavedList
 
 data class SavedListDetails(
     val savedList: SavedList,
-    val pagedItems: PagedResult<SavableMovie>,
+    val pagedItems: PagedResult<SavedMovie>
 )
