--- domain/src/main/java/com/baghdad/domain/model/search/SearchFilter.kt
@@ -1,10 +1,0 @@
-package com.baghdad.domain.model.search
-
-import com.baghdad.entity.media.Genre
-
-data class SearchFilter(
-    val minimumYear: Int?,
-    val maximumYear: Int?,
-    val minimumRating: Int,
-    val selectedGenres: List<Genre>
-)
