+++ entity/src/main/java/com/baghdad/entity/savedList/SavedList.kt
@@ -3,5 +3,5 @@ package com.baghdad.entity.savedList
 data class SavedList(
     val id: Long,
     val name: String,
-    val itemCount: Int,
+    val itemCount: Int
 )
